pyproject.toml
[tool.poetry]
name = "nsp-stc"
version = "0.1.0"
description = ""
authors = ["Philip Huang <p208p2002@gmail.com>"]
readme = "README.md"
packages = [{include = "zhpr"}]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.5.1"
openpyxl = "^3.0.10"
scikit-learn = "^1.1.3"
transformers = "^4.24.0"
seqeval = "^1.2.2"


[tool.poetry.group.dev.dependencies]
poethepoet = "^0.16.4"
light-the-torch = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
# https://github.com/nat-n/poethepoet
setup-env = "sh setup-lightning.sh" # simple command based task
